requirements.txt
dj-database-url==0.5.0
Django==2.2.1
django-crispy-forms==1.7.2
django-heroku==0.3.1
files==1.1.1
gunicorn==19.9.0
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
sqlparse==0.3.0
whitenoise==4.1.2
